// types/Country.ts
export type Country = {
  name: string;
  alpha2Code: string;
  capital: string;
  altSpellings: string[];
  subregion: string;
  region: string;
  population: number;
  latlng: number[];
  demonym: string;
  borders: string[];
  numericCode: string;
  flags: {
    svg: string;
    png: string;
  };
  translations: {
    br: string;
    pt: string;
    nl: string;
    hr: string;
    fa: string;
    de: string;
    es: string;
    fr: string;
    ja: string;
    it: string;
    hu: string;
  };
  flag: string;
};